Configuration - /Bifrost/Application geneartion up and running
