+++ Source/Bifrost.Web/Configuration/net4x/ConfigurationRouteHttpHandler.cs
@@ -3,32 +3,31 @@
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 using System;
-using System.Diagnostics;
-using System.Text;
 using System.Web;
 using Bifrost.Configuration;
-using Bifrost.Web.Assets;
-using Bifrost.Web.Proxies;
-using Newtonsoft.Json;
 
 namespace Bifrost.Web.Configuration
 {
     public class ConfigurationRouteHttpHandler : IHttpHandler
     {
-        string _configurationAsString;
-
-        public bool IsReusable { get { return true; } }
-
         WebConfiguration _webConfiguration;
+        ConfigurationAsJavaScript _configurationAsJavaScript;
 
-        public ConfigurationRouteHttpHandler() : this(Configure.Instance.Container.Get<WebConfiguration>())
+        public ConfigurationRouteHttpHandler() : 
+            this(
+                Configure.Instance.Container.Get<WebConfiguration>(),
+                Configure.Instance.Container.Get<ConfigurationAsJavaScript>())
         {
         }
 
-        public ConfigurationRouteHttpHandler(WebConfiguration webConfiguration)
+        public ConfigurationRouteHttpHandler(WebConfiguration webConfiguration, ConfigurationAsJavaScript configurationAsJavaScript)
         {
             _webConfiguration = webConfiguration;
+            _configurationAsJavaScript = configurationAsJavaScript;
         }
+
+        public bool IsReusable => true;
+
 
         public void ProcessRequest(HttpContext context)
         {
@@ -64,68 +63,8 @@ namespace Bifrost.Web.Configuration
 
         void OutputContent(HttpContext context)
         {
-            InitializeIfNotInitialized();
             context.Response.ContentType = "text/javascript";
-            context.Response.Write(_configurationAsString);
-        }
-
-
-        string GetResource(string name)
-        {
-            try
-            {
-                var stream = typeof(ConfigurationRouteHttpHandler).Assembly.GetManifestResourceStream(name);
-                var bytes = new byte[stream.Length];
-                stream.Read(bytes, 0, bytes.Length);
-                var content = UTF8Encoding.UTF8.GetString(bytes);
-                return content;
-            } catch
-            {
-                Debug.Write($"Couldn't get the resource '{name}'");
-                throw;
-            }
-        }
-
-        void InitializeIfNotInitialized()
-        {
-            if (!string.IsNullOrEmpty(_configurationAsString)) return;
-
-            var proxies = Configure.Instance.Container.Get<GeneratedProxies>();
-            
-            var assetsManager = Configure.Instance.Container.Get<IAssetsManager>();
-
-            var builder = new StringBuilder();
-
-            if (_webConfiguration.ScriptsToInclude.JQuery)
-                builder.Append(GetResource("Bifrost.Web.Scripts.jquery-2.1.3.min.js"));
-
-            if (_webConfiguration.ScriptsToInclude.Knockout)
-                builder.Append(GetResource("Bifrost.Web.Scripts.knockout-3.2.0.debug.js"));
-
-            if (_webConfiguration.ScriptsToInclude.SignalR)
-                builder.Append(GetResource("Bifrost.Web.Scripts.jquery.signalR-2.2.0.js"));
-
-            if (_webConfiguration.ScriptsToInclude.JQueryHistory)
-                builder.Append(GetResource("Bifrost.Web.Scripts.jquery.history.js"));
-
-            if (_webConfiguration.ScriptsToInclude.Require)
-            {
-                builder.Append(GetResource("Bifrost.Web.Scripts.require.js"));
-                builder.Append(GetResource("Bifrost.Web.Scripts.order.js"));
-                builder.Append(GetResource("Bifrost.Web.Scripts.text.js"));
-                builder.Append(GetResource("Bifrost.Web.Scripts.domReady.js"));
-                builder.Append(GetResource("Bifrost.Web.Scripts.noext.js"));
-            }
-
-            if (_webConfiguration.ScriptsToInclude.Bifrost)
-                builder.Append(GetResource("Bifrost.Web.Scripts.Bifrost.debug.js"));
-
-            builder.Append(proxies.All);
-
-            var files = assetsManager.GetFilesForExtension("js");
-            var serialized = JsonConvert.SerializeObject(files);
-            builder.AppendFormat("Bifrost.assetsManager.initializeFromAssets({0});", serialized);
-            _configurationAsString = builder.ToString();
+            context.Response.Write(_configurationAsJavaScript.AsString);
         }
     }
 }
